# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cultural Site API
  version: 1.0.0
  description: API for managing cultural sites, proposals, reviews, and user data.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user.
        username:
          type: string
          maxLength: 20
          description: User's chosen username.
        email:
          type: string
          format: email
          description: User's email address.
        googleId:
          type: string
          description: Google ID if authenticated via Google.
        profileImage:
          type: string
          format: url
          description: URL to the user's profile image.
        role:
          type: string
          enum: ['user', 'admin']
          default: 'admin'
          description: User's role in the system. temporarily default value is 'admin' to demonstrate the functions.
        active:
          type: boolean
          default: true
          description: Field for soft deletion.
        favoriteSites:
          type: array
          items:
            type: string
            description: ObjectId of a CulturalSite.
          description: List of cultural sites marked as favorites by the user.
        bio:
          type: string
          maxLength: 200
          description: User's short biography.
      required:
        - email
        - googleId
...
paths:
...
  /auth/google/callback:
    get:
      summary: Google authentication callback
      tags:
        - Authentication
      operationId: googleAuthCallback
      description: Handles the callback from Google after authentication. Issues a JWT on success.
      responses:
        302:
          description: Redirects to frontend on success or failure.
        200:
          description: Authentication successful, JWT issued (though typically a redirect happens).
        401:
          description: Google authentication failed.
...
  /users/updateRole/{userId}:
    patch:
      summary: Update a user's role (Admin only)
      tags:
        - Users
      operationId: updateUserRole
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: Unique ID of the user whose role is to be updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: ['user', 'admin']
              required:
                - role
            examples:
              changeToAdmin:
                value:
                  role: "admin"
      responses:
        200:
          description: User role updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid input.
        401:
          description: Unauthorized.
        403:
          description: Forbidden, only admins can update user roles.
        404:
          description: User not found.
        500:
          description: Internal server error.
...